refactor(workout-plan): extract estimates cluster to workout-plan-estimates.js (WP3.4c)
Behavior-preserving move-only split (WP3.4c of Refactor Plan v3 Phase 3).
Moves the Workout Controls profile-estimate / learned-calibration /
fatigue-context / trace / nudge cluster (25 functions + weightUserDirty and
latestTracePayload module state) out of workout-plan.js into the new
static/js/modules/workout-plan-estimates.js.

The cluster is self-contained: it imports only toast.js, fetch-wrapper.js
(api), and pure helpers from workout-plan-helpers.js — nothing from the
monolith — so this is a plain one-directional import, no callback injection
needed. The monolith imports the three entry points
(applyUserProfileEstimateForSelectedExercise, initializeWeightDirtyTracking,
bindEstimateTraceToggle) and re-exports the first to preserve its export
surface. The nine estimate-only helper imports are pruned from the monolith.

The weight-dirty flag now lives in the estimates module, so the pre-existing
private resetWeightDirtyForTests() (already used in production, misnamed) is
renamed to an exported resetWeightUserDirty() and the two monolith call sites
(handleExerciseSelection, resetFormFields) call it instead of assigning the
bare variable. Moved bodies are byte-identical to HEAD apart from the three
added export keywords and that one rename.

No Python/DB/API/response-shape/CSS/CI-context change. Vitest 88/88;
E2E workout-plan + exercise-interactions + fatigue-context + superset-edge-cases
+ replace-exercise-errors + learned-calibration + user-profile 108 passed
(learned-calibration:506 combined-run contamination flake, 8/8 in isolation).
